// src/generated/kubb/zod/errorStreamingEventSchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import { errorPayloadSchema } from "./errorPayloadSchema.ts";
import { z } from "zod";

/**
 * @description A streaming event that indicated an error was emitted.
 */
export const errorStreamingEventSchema = z
  .object({
    type: z
      .enum(["error"])
      .default("error")
      .describe("The type of the event, always `error`."),
    sequence_number: z
      .number()
      .int()
      .describe("The sequence number of the event that was emitted."),
    error: z.lazy(() => errorPayloadSchema).and(z.any()),
  })
  .describe("A streaming event that indicated an error was emitted.");