// src/generated/kubb/zod/responseCompletedStreamingEventSchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import { responseResourceSchema } from "./responseResourceSchema.ts";
import { z } from "zod";

/**
 * @description A streaming event that indicated the response was completed.
 */
export const responseCompletedStreamingEventSchema = z
  .object({
    type: z
      .enum(["response.completed"])
      .default("response.completed")
      .describe("The type of the event, always `response.completed`."),
    sequence_number: z
      .number()
      .int()
      .describe("The sequence number of the event that was emitted."),
    response: z.lazy(() => responseResourceSchema).and(z.any()),
  })
  .describe("A streaming event that indicated the response was completed.");