// src/generated/kubb/zod/responseContentPartAddedStreamingEventSchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import { inputFileContentSchema } from "./inputFileContentSchema.ts";
import { inputImageContentSchema } from "./inputImageContentSchema.ts";
import { inputTextContentSchema } from "./inputTextContentSchema.ts";
import { outputTextContentSchema } from "./outputTextContentSchema.ts";
import { reasoningTextContentSchema } from "./reasoningTextContentSchema.ts";
import { refusalContentSchema } from "./refusalContentSchema.ts";
import { summaryTextContentSchema } from "./summaryTextContentSchema.ts";
import { textContentSchema } from "./textContentSchema.ts";
import { z } from "zod";

/**
 * @description A streaming event that indicated a content part was added.
 */
export const responseContentPartAddedStreamingEventSchema = z
  .object({
    type: z
      .enum(["response.content_part.added"])
      .default("response.content_part.added")
      .describe("The type of the event, always `response.content_part.added`."),
    sequence_number: z
      .number()
      .int()
      .describe("The sequence number of the event that was emitted."),
    item_id: z.string().describe("The ID of the item that was updated."),
    output_index: z
      .number()
      .int()
      .describe("The index of the output item that was updated."),
    content_index: z
      .number()
      .int()
      .describe("The index of the content part that was added."),
    part: z
      .union([
        z.lazy(() => inputTextContentSchema),
        z.lazy(() => outputTextContentSchema),
        z.lazy(() => textContentSchema),
        z.lazy(() => summaryTextContentSchema),
        z.lazy(() => reasoningTextContentSchema),
        z.lazy(() => refusalContentSchema),
        z.lazy(() => inputImageContentSchema),
        z.lazy(() => inputFileContentSchema),
      ])
      .describe("A content part that makes up an input or output item."),
  })
  .describe("A streaming event that indicated a content part was added.");