// src/generated/kubb/zod/responseCreatedStreamingEventSchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import { responseResourceSchema } from "./responseResourceSchema.ts";
import { z } from "zod";

/**
 * @description A streaming event that indicated the response was created.
 */
export const responseCreatedStreamingEventSchema = z
  .object({
    type: z
      .enum(["response.created"])
      .default("response.created")
      .describe("The type of the event, always `response.created`."),
    sequence_number: z
      .number()
      .int()
      .describe("The sequence number of the event that was emitted."),
    response: z.lazy(() => responseResourceSchema).and(z.any()),
  })
  .describe("A streaming event that indicated the response was created.");