// src/generated/kubb/zod/responseFailedStreamingEventSchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import { responseResourceSchema } from "./responseResourceSchema.ts";
import { z } from "zod";

/**
 * @description A streaming event that indicated the response had failed.
 */
export const responseFailedStreamingEventSchema = z
  .object({
    type: z
      .enum(["response.failed"])
      .default("response.failed")
      .describe("The type of the event, always `response.failed`."),
    sequence_number: z
      .number()
      .int()
      .describe("The sequence number of the event that was emitted."),
    response: z.lazy(() => responseResourceSchema).and(z.any()),
  })
  .describe("A streaming event that indicated the response had failed.");