// src/generated/kubb/zod/responseFunctionCallArgumentsDeltaStreamingEventSchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import { z } from "zod";

/**
 * @description A streaming event that indicated function call arguments were incrementally added.
 */
export const responseFunctionCallArgumentsDeltaStreamingEventSchema = z
  .object({
    type: z
      .enum(["response.function_call_arguments.delta"])
      .default("response.function_call_arguments.delta")
      .describe(
        "The type of the event, always `response.function_call_arguments.delta`.",
      ),
    sequence_number: z
      .number()
      .int()
      .describe("The sequence number of the event that was emitted."),
    item_id: z
      .string()
      .describe("The ID of the tool call item that was updated."),
    output_index: z
      .number()
      .int()
      .describe("The index of the output item that was updated."),
    delta: z.string().describe("The arguments delta that was appended."),
    obfuscation: z.optional(
      z
        .string()
        .describe(
          "An obfuscation string that was added to pad the event payload.",
        ),
    ),
  })
  .describe(
    "A streaming event that indicated function call arguments were incrementally added.",
  );