// src/generated/kubb/zod/responseFunctionCallArgumentsDoneStreamingEventSchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import { z } from "zod";

/**
 * @description A streaming event that indicated function call arguments were completed.
 */
export const responseFunctionCallArgumentsDoneStreamingEventSchema = z
  .object({
    type: z
      .enum(["response.function_call_arguments.done"])
      .default("response.function_call_arguments.done")
      .describe(
        "The type of the event, always `response.function_call_arguments.done`.",
      ),
    sequence_number: z
      .number()
      .int()
      .describe("The sequence number of the event that was emitted."),
    item_id: z
      .string()
      .describe("The ID of the tool call item that was updated."),
    output_index: z
      .number()
      .int()
      .describe("The index of the output item that was updated."),
    arguments: z
      .string()
      .describe("The final arguments string that was emitted."),
  })
  .describe(
    "A streaming event that indicated function call arguments were completed.",
  );