// src/generated/kubb/zod/responseInProgressStreamingEventSchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import { responseResourceSchema } from "./responseResourceSchema.ts";
import { z } from "zod";

/**
 * @description A streaming event that indicated the response was in progress.
 */
export const responseInProgressStreamingEventSchema = z
  .object({
    type: z
      .enum(["response.in_progress"])
      .default("response.in_progress")
      .describe("The type of the event, always `response.in_progress`."),
    sequence_number: z
      .number()
      .int()
      .describe("The sequence number of the event that was emitted."),
    response: z.lazy(() => responseResourceSchema).and(z.any()),
  })
  .describe("A streaming event that indicated the response was in progress.");