// src/generated/kubb/zod/responseOutputItemAddedStreamingEventSchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import { itemFieldSchema } from "./itemFieldSchema.ts";
import { z } from "zod";

/**
 * @description A streaming event that indicated an output item was added to the response.
 */
export const responseOutputItemAddedStreamingEventSchema = z
  .object({
    type: z
      .enum(["response.output_item.added"])
      .default("response.output_item.added")
      .describe("The type of the event, always `response.output_item.added`."),
    sequence_number: z
      .number()
      .int()
      .describe("The sequence number of the event that was emitted."),
    output_index: z
      .number()
      .int()
      .describe("The index of the output item that was added."),
    item: z.union([z.lazy(() => itemFieldSchema).and(z.any()), z.null()]),
  })
  .describe(
    "A streaming event that indicated an output item was added to the response.",
  );