// src/generated/kubb/zod/responseOutputItemDoneStreamingEventSchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import { itemFieldSchema } from "./itemFieldSchema.ts";
import { z } from "zod";

/**
 * @description A streaming event that indicated an output item was completed.
 */
export const responseOutputItemDoneStreamingEventSchema = z
  .object({
    type: z
      .enum(["response.output_item.done"])
      .default("response.output_item.done")
      .describe("The type of the event, always `response.output_item.done`."),
    sequence_number: z
      .number()
      .int()
      .describe("The sequence number of the event that was emitted."),
    output_index: z
      .number()
      .int()
      .describe("The index of the output item that was completed."),
    item: z.union([z.lazy(() => itemFieldSchema).and(z.any()), z.null()]),
  })
  .describe("A streaming event that indicated an output item was completed.");